3rd Place VR (hello world) Hackathon
Tried to make a 3D rendering of the stars orbiting a 3D text of "hello world" while moving at the speed of light. 

Turns out javascript is crazy to do this. Yet I will continue to try and implement more to see where I can go. 

I have a 2D plane to have an illusion of 3D using the z axis on 2D shapes.